// src/model/contract-info.d.ts
export interface ContractInfo {
  codeContract: string;//合同编号
  nameContract: string;//合同名称
  jeContract: string;//合同标的（审计）额
  projTypeContract: string;//项目性质
  projPartNameContract: string;//项目单元名称
  projSettTypeContract: string;//项目核算类别
  paySeasonContract: string;//付款原因
  readyPayContract: string;//已付款额度
  nowApplyContract: string;//本次申请金额
  payDeptContract: string;//付款单位名称
  getDeptContract: string;//收款单位名称
  applyTimeContract: string;//申请时间
  applyContract: string;//申请人
}